Test two decimals water measurement

+++ database/migrations/2024_06_22_074135_create_measurements_table.php
@@ -24,9 +24,9 @@ return new class extends Migration
             $table->foreign('meter_id')->references('id')->on('meters')
                 ->onDelete('cascade')->onUpdate('cascade');
 
-            $table->integer('start_value');
-            $table->integer('end_value');
-            $table->integer('difference');
+            $table->decimal('start_value',10,2);
+            $table->decimal('end_value',10,2);
+            $table->decimal('difference',10,2);
             $table->date('date');
             $table->timestamps();
         });
